Updates docs to include a way for the server to redirect to callback
There is no where in the docs that specifies how to redirect a user from the server. This is important because there are times where a server component checks if a user is authenticated and needs to redirect them to the sign-in page if they are not. This addition in the docs will let developer know that they can specify a callbackUrl that will redirect the user to the specified page instead of the root page. This is crucial for user experience as they can be redirected back to the page they came from.